Simplify renaming of initial bucket
This change puts bucket name inline so that 'default' bucket
is just the synonym for initial bucket, but it's actual name is an attribute
so we can do things like:

 --set couchbaseCluster.buckets.default.name=travel-sample

Btw, buckets are maps because of the  way helm does overrides via --set and -f.
Maps make it easy to access and change specific bucket attrs.