K8S-2732: Manually define RBAC resources
In the past Users were created and automatically bound to roles for
convience. This came at the cost of customization now that
scopes/collections along with LDAP allow for much more advanced
configurations.

This change allows users to be set with an 'autobind' property that
can disable the automatic resource creation and manually create
themselves.